Moved splash
